pyproject.toml
[project]
name = "django-dev"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.4",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.0",
    "google-adk[a2a]>=1.8.0",
    "mysqlclient>=2.2.7",
    "requests>=2.32.4",
]
